backend/requirements.txt
openai-agents
fastapi
uvicorn
qdrant-client
google-genai
python-dotenv
python-frontmatter
openai-chatkit
supabase
pytest
pytest-asyncio